Refactor and rename test to check for url's health
