+++ tox.ini
@@ -16,12 +16,12 @@ commands =
     py27: find {toxinidir}/ -name "*.pyc" -type f -delete
 
 # Check URLs health status
-[testenv:urls_check]
+[testenv:urls_check_health]
 commands =
     pip install pipenv
     pipenv install --dev
     pipenv run python setup.py install
-    pipenv run pytest --cov=dbcollection {toxinidir}/dbcollection/tests/datasets/__test_check_urls.py
+    pipenv run pytest --cov=dbcollection {toxinidir}/dbcollection/tests/datasets/__test_check_urls_health.py
 
 [testenv:flake8]
 commands =
